Improvements To Help With Multiple Executions In The Same Maven Run:
AbstractThriftMojo: Commented-out call to FileUtils#cleanDirectory - this was preventing multiple executions of the plugin with varying config during the same maven goal execution.

org.apache.thrift.maven.Thrift: Commented-out call to #moveGeneratedFiles. This method aims to omit the parent generation directory created by the thrift compiler (e.g. gen-java, gen-cpp). This is not compatible with the aim of retaining the generation output of multiple different generator configurations.

TestThrift: Modifications to reflect the above changes. Also tweaks to help with system PATH sensitivity and to output the actual thrift command line that was executed if an error occurred.